리펙토링 - 컴포넌트 라우팅을 위해 디렉토리 구조를 수정하였다. - 달력과 네비의 디자인을 수정하였다. - dom 초기 렌더 시 라우팅 제어 수정
- 주소로 접근이 안되는 문제가 발생하였다. 해결요망.
