+++ .eslintrc
@@ -5,7 +5,9 @@
     "import/named": 0,
     "import/no-unresolved": 0,
     "import/prefer-default-export": 0,
-    "camelcase": 0
+    "camelcase": 0,
+    "no-restricted-syntax": 0,
+    "radix": 0,
   },
   "parserOptions": {
     "sourceType": "module"
