+++ client/src/App/components/calendar/calendar.scss
@@ -1,7 +1,10 @@
 $border: 2px solid rgba($color: #000000, $alpha: .1);
 
 .calendar {
+  display: flex;
+  flex-direction: column;
   .week {
+    margin: auto;
     display: flex;
     .day {
       width: 8rem;
