+++ client/src/App/components/dailyHistory/caseHistory/caseHistory.scss
