+++ client/src/App/components/dailyHistory/caseHistory/index.js
