+++ client/src/App/components/dailyHistory/dailyHistory.scss
