+++ client/src/App/components/hkbForm/hkbForm.scss
@@ -1,7 +1,6 @@
 .hkb_form {
   background-color: #fcfcfd;
   border-radius: 10px;
-  margin-top: 15px;
   padding: 10px;
 
   .row {
