+++ client/src/App/components/layout/container/index.js
@@ -3,8 +3,10 @@ import Navigation from './navigation';
 
 const Container = (child) => (
   `<div class="container">
-      ${Navigation()}
+    ${Navigation()}
+    <div class="contents">
       ${child}
+    </div>
   </div>`
 );
