+++ client/src/App/components/layout/container/navigation/navigation.scss
@@ -34,7 +34,7 @@
     align-items: center;
     justify-content: center;
     border-radius: 20px;
-    margin-top: 15px;
+    margin: 15px 0;
     padding: 8px 0;
 
     .bar {
