+++ client/src/App/pages/calendar/index.js
@@ -1,10 +1,10 @@
-import { Layout, Calendar } from '../../components';
+import { Calendar } from '../../components';
 import { getState } from '../../store';
 
 const CalendarPage = () => {
   const currentMonth = getState('currentMonth');
 
-  return Layout(`<div class="calendar_page">${Calendar(currentMonth)}</div>`);
+  return (`<div class="calendar_page">${Calendar(currentMonth)}</div>`);
 };
 
 export default CalendarPage;
