+++ client/src/App/pages/graph/index.js
@@ -1,0 +1,5 @@
+const Graph = () => (
+  '<div class="graph">통계 페이지</div>'
+);
+
+export default Graph;
