--- client/src/pages/graph/index.js
@@ -1,7 +1,0 @@
-import { Layout } from '../../components';
-
-const Graph = () => (
-  Layout('<div class="graph">통계 페이지</div>')
-);
-
-export default Graph;
